display muliply images on 1 spot

+++ frontend/src/components/bookings/BookingIndexItem.js
@@ -7,6 +7,7 @@ const BookingIndexItem = ({ booking, spot }) => {
 
   const handleDelete = async e => {
     e.preventDefault();
+    if (!window.confirm('Do you want to delete this booking?')) return;
     await dispatch(deleteBooking(booking.id));
   };
